use workaround and take new snaps
