+++ src/lib.rs
@@ -264,3 +264,4 @@ pub mod scc;
 pub mod string;
 pub mod symbol;
 pub mod token;
+pub mod utils;
